ca-slack-demo/pyproject.toml
[project]
name = "slack-analytics-agents"
version = "0.1.0"
description = "Multi-agent Slack bot for Conversational Analytics (CA API)"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-geminidataanalytics>=0.1.0",
    "slack-bolt>=1.18.0",
    "vl-convert-python>=1.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "mcp>=1.27.0",
    "gspread>=6.2.1",
    "google-auth>=2.49.1",
    "google-api-python-client>=2.193.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "integration: requires real GCP credentials and CA agent IDs",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
